// dmgcc/02_WebFrontEnd/src/saga/Types/SLA.Types.ts
export enum SLAActionTypes {
    TERMS_CONDITION_UPLOAD_REQUEST = 'TERMS_CONDITION_UPLOAD_REQUEST',
    TERMS_CONDITION_UPLOAD_SUCCESS = 'TERMS_CONDITION_UPLOAD_SUCCESS',
    TERMS_CONDITION_UPLOAD_ERROR = 'TERMS_CONDITION_UPLOAD_ERROR',
    SLA_UPLOAD_REQUEST = 'SLA_UPLOAD_REQUEST',
    SLA_UPLOAD_SUCCESS = 'SLA_UPLOAD_SUCCESS',
    SLA_UPLOAD_ERROR = 'SLA_UPLOAD_ERROR',
    BILLING_CYCLE_DROPDOWN_REQUEST = 'BILLING_CYCLE_DROPDOWN_REQUEST',
    BILLING_CYCLE_DROPDOWN_SUCCESS = 'BILLING_CYCLE_DROPDOWN_SUCCESS',
    BILLING_CYCLE_DROPDOWN_ERROR = 'BILLING_CYCLE_DROPDOWN_ERROR',
    GET_PER_INVOICE_REQUEST = 'GET_PER_INVOICE_REQUEST',
    GET_PER_INVOICE_SUCCESS = 'GET_PER_INVOICE_SUCCESS',
    GET_PER_INVOICE_ERROR = 'GET_PER_INVOICE_ERROR',
    POST_PER_INVOICE_REQUEST = 'POST_PER_INVOICE_REQUEST',
    POST_PER_INVOICE_SUCCESS = 'POST_PER_INVOICE_SUCCESS',
    POST_PER_INVOICE_ERROR = 'POST_PER_INVOICE_ERROR',
    GET_PROJECT_SLA_DETAILS_REQUEST = 'GET_PROJECT_SLA_DETAILS_REQUEST',
    GET_PROJECT_SLA_DETAILS_SUCCESS = 'GET_PROJECT_SLA_DETAILS_SUCCESS',
    GET_PROJECT_SLA_DETAILS_ERROR = 'GET_PROJECT_SLA_DETAILS_ERROR',
    GET_ORGANIZATION_REQUEST = 'GET_ORGANIZATION_REQUEST',
    GET_ORGANIZATION_SUCCESS = 'GET_ORGANIZATION_SUCCESS',
    GET_ORGANIZATION_ERROR = 'GET_ORGANIZATION_ERROR',
    GET_BILLING_CYCLE_REQUEST = 'GET_BILLING_CYCLE_REQUEST',
    GET_BILLING_CYCLE_SUCCESS = 'GET_BILLING_CYCLE_SUCCESS',
    GET_BILLING_CYCLE_ERROR = 'GET_BILLING_CYCLE_ERROR',
    GET_MATERIAL_DESCRIPTION_REQUEST = 'GET_MATERIAL_DESCRIPTION_REQUEST',
    GET_MATERIAL_DESCRIPTION_SUCCESS = 'GET_MATERIAL_DESCRIPTION_SUCCESS',
    GET_MATERIAL_DESCRIPTION_ERROR = 'GET_MATERIAL_DESCRIPTION_ERROR',
    GET_COUNTRY_LIST_REQUEST = 'GET_COUNTRY_LIST_REQUEST',
    GET_COUNTRY_LIST_SUCCESS = 'GET_COUNTRY_LIST_SUCCESS',
    GET_COUNTRY_LIST_ERROR = 'GET_COUNTRY_LIST_ERROR',
    GET_CONTRACT_OPITION_REQUEST = 'GET_CONTRACT_OPITION_REQUEST',
    GET_CONTRACT_OPITION_SUCCESS = 'GET_CONTRACT_OPITION_SUCCESS',
    GET_CONTRACT_OPITION_ERROR = 'GET_CONTRACT_OPITION_ERROR',
    GET_CURRENCY_LIST_REQUEST = 'GET_CURRENCY_LIST_REQUEST',
    GET_CURRENCY_LIST_SUCCESS = 'GET_CURRENCY_LIST_SUCCESS',
    GET_CURRENCY_LIST_ERROR = 'GET_CURRENCY_LIST_ERROR',
    GET_TARIFF_SHEET_MATERIAL_DESCRIPTION_REQUEST = 'GET_TARIFF_SHEET_MATERIAL_DESCRIPTION_REQUEST',
    GET_TARIFF_SHEET_MATERIAL_DESCRIPTION_SUCCESS = 'GET_TARIFF_SHEET_MATERIAL_DESCRIPTION_SUCCESS',
    GET_TARIFF_SHEET_MATERIAL_DESCRIPTION_ERROR = 'GET_TARIFF_SHEET_MATERIAL_DESCRIPTION_ERROR',
    
   SENT_MAIL_SLA_REQUEST = 'SENT_MAIL_SLA_REQUEST',
   SENT_MAIL_SLA_SUCCESS = 'SENT_MAIL_SLA_SUCCESS',
   SENT_MAIL_SLA_ERROR = 'SENT_MAIL_SLA_ERROR',
}